12 - 17
- type: Add Practice Code
- body
class Token
Member
Token* down_Token, Token* up_Token
These represent the correlation of each token.
map[n][m][0] represents side's color.
map[n][m][1] -1 indicates that there is no token. If the index is for a token, it represents the index of the token at the top of all tokens in that column.